Move WebhookRestServer operator/webhooks
